+++ changes.md
@@ -4,6 +4,7 @@
 1. Try to provide workaround for websites where requirejs already loaded.
    This makes problem by direct loading of jquery and jquery-ui 
 2. Provide workaround for older version of jquery-ui 
+3. Prompt for input of command arguments
 
 
 ## Changes in 3.5